Move argument types to separate file
